voice_input/requirements.txt
python-dotenv
sounddevice
numpy
websocket-client
pynput
pyperclip
loguru
